requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
scikit-learn==1.5.1
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2
joblib==1.4.2
pydantic==2.7.4
PyYAML==6.0.1
mangum==0.17.0
python-dotenv==1.0.1
requests==2.32.3
pytest==8.2.2
httpx==0.27.0
google-generativeai==0.7.2
cachetools==5.5.0 